+++ src/main/kotlin/app/revanced/library/adb/Commands.kt
@@ -5,8 +5,10 @@ import se.vidstige.jadb.RemoteFile
 import se.vidstige.jadb.ShellProcessBuilder
 import java.io.File
 
-
-internal fun JadbDevice.buildCommand(command: String, su: Boolean = true): ShellProcessBuilder {
+internal fun JadbDevice.buildCommand(
+    command: String,
+    su: Boolean = true,
+): ShellProcessBuilder {
     if (su) return shellProcessBuilder("su -c \'$command\'")
 
     val args = command.split(" ") as ArrayList<String>
@@ -15,14 +17,19 @@ internal fun JadbDevice.buildCommand(command: String, su: Boolean = true): Shell
     return shellProcessBuilder(cmd, *args.toTypedArray())
 }
 
-internal fun JadbDevice.run(command: String, su: Boolean = true) =
-    this.buildCommand(command, su).start()
+internal fun JadbDevice.run(
+    command: String,
+    su: Boolean = true,
+) = this.buildCommand(command, su).start()
 
-internal fun JadbDevice.hasSu() =
-    this.run("whoami", true).waitFor() == 0
+internal fun JadbDevice.hasSu() = this.run("whoami", true).waitFor() == 0
 
-internal fun JadbDevice.push(file: File, targetFilePath: String) =
-    push(file, RemoteFile(targetFilePath))
+internal fun JadbDevice.push(
+    file: File,
+    targetFilePath: String,
+) = push(file, RemoteFile(targetFilePath))
 
-internal fun JadbDevice.createFile(targetFile: String, content: String) =
-    push(content.byteInputStream(), System.currentTimeMillis(), 644, RemoteFile(targetFile))
+internal fun JadbDevice.createFile(
+    targetFile: String,
+    content: String,
+) = push(content.byteInputStream(), System.currentTimeMillis(), 644, RemoteFile(targetFile))